requirements.txt
PyQt6>=6.0.0
Pillow>=9.0.0

